refactoring APIs
