--- modules/back-end/src/Application/Services/IUsageAppService.cs
@@ -1,8 +1,0 @@
-using Application.Usages;
-
-namespace Application.Services;
-
-public interface IUsageAppService
-{
-    Task SaveRecordsAsync(AggregatedUsageRecords records);
-}
