+++ modules/back-end/src/Application/Services/IWorkspaceService.cs
@@ -1,3 +1,4 @@
+using Application.Usages;
 using Domain.Workspaces;
 
 namespace Application.Services;
@@ -8,5 +9,7 @@ public interface IWorkspaceService : IService<Workspace>
 
     Task<string> GetDefaultWorkspaceAsync();
 
-    Task<int> GetUsageAsync(Guid workspaceId, string feature);
+    Task<int> GetFeatureUsageAsync(Guid workspaceId, string feature);
+
+    Task SaveRecordsAsync(AggregatedUsageRecords records);
 }
